Implement Favorite Locations
Users can save locations to access later.
Clicking the favorite icon saves the Location objects
(city, state, latitude, and longitude) as FavoriteLocations[].
These can be found in the favorites menu.
The favorite icon serves as a favorites toggle, clicking it
on a saved location will remove it from the favorites array.